requirements.txt
altair==4.2.0
numpy==1.23.4
pandas==1.5.1
plotly==5.13.0
snowflake_connector_python==3.0.0
streamlit==1.18.0
snowflake-snowpark-python